Refactor: Enforce interface locality by moving DTO types to API clients
This resolves the interface-locality issues where API response/request interfaces that are only used in a single file were centrally stored in the `src/types` directory.

- Moved single-use API request/response types (e.g. `MapBootstrapResponse`, `UploadResponse`, `ReviewCreateRequest`, etc.) from `src/types/*` to their respective API client modules in `src/api/*Client.ts`.
- Updated all import paths and internal barrel exports to fix typecheck constraints.
- Retained completely unused types like `DiscoverySearchResponse` to prevent regressions.
- Kept the `package-lock.json` untouched regarding peer dependencies to minimize noise.
- Cleaned up the `test/unit/interface-locality-source-quality.test.ts` baseline assertion safely to ensure barrel import health is strictly guarded.

All tests, linter, and typechecks pass.
